ttm/radeon/nouveau: Check the DMA address from TTM against known value.
.. instead of checking against the DMA_ERROR_CODE value which is
per-platform specific. The zero value is a known invalid value
that the TTM layer sets on the dma_address array if it is not
used. We can't use pci_dma_mapping_error as that is IOMMU
specific (some check for a specific physical address, some
for ranges, some just do a check against zero).

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>