ttm: Pass in 'struct device' to TTM so it can do DMA API on behalf of device.
We want to pass in the 'struct device' to the TTM layer in a nice way.
This makes accounting of pages and their DMA addresses correct when using
CONFIG_DMA_API_DEBUG=y. Visually it means that when using
'debug_dma_dump_mappings()' it will print the name of the device
(say 'radeon') instead of 'fallback'. On PPC64, using 'struct dev *' instead
of 'NULL' is required - otherwise it BUGs() on bootup.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>